requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
openai
jinja2
resend
python-dotenv
gunicorn
python-jose[cryptography]
pytesseract
Pillow
python-magic
pypdf
google-generativeai
openpyxl
pandas
weasyprint
supabase==2.13.0
celery==5.3.4
redis==5.0.1
sentry-sdk[fastapi]==1.39.1
